Move list.
