+++ crates/cli/src/commands/plugin/mod.rs
@@ -1,5 +1,7 @@
 mod add;
+mod list;
 mod remove;
 
 pub use add::*;
+pub use list::*;
 pub use remove::*;
